feat: move blogs from home page to blog page
