+++ bootcamp-milestone-3/src/app/page.tsx
@@ -1,32 +1,4 @@
-import BlogPreview from "@/components/blogPreview";
-import connectDB  from "../database/database";
-import Blog from "@/database/blogSchema";
-
-async function getBlogs() {
-  const url: string = process.env.MONGO_URI as string;
-
-  await connectDB();
-
-  try {
-    const blogs = await Blog.find().sort({ date: -1 }).orFail();
-
-    const formattedBlogs = blogs.map((blog) => ({
-      title: blog.title,
-      content: blog.content,
-      description: blog.description,
-      slug: blog.slug,
-      date: blog.date.toString(),
-    }));
-
-    return formattedBlogs;
-  } catch (err) {
-    return null;
-  }
-}
-
-export default async function Home() {
-  const blogs = await getBlogs();
-
+export default function Home() {
   return (
     <div>
       <div className="home-border-wrap">
@@ -41,12 +13,6 @@ export default async function Home() {
                 src="./images/meSteezTransparent.png"
                 width="700"
               />
-            </div>
-
-            <div id="blog-container">
-              {blogs.map((blog) => (
-                <BlogPreview {...blog} />
-              ))}
             </div>
           </div>
         </main>
